debian: create home in cephadm postinst
The code for creating the user only runs when the user does not yet exist, so it only affects systems where the user is not manually managed by the operator.
Creation of the SSH directory for SSH keys will fail if the user home does not yet exist.
So the only way that this will ever work is if the operator (or some other part of the workflow) either *always* creates that user or *always* creates the home directory (without the user).
If the package is installed without the home directory or user existing it will fail since the *mkdir* (as it is called without `-p`) will fail with `ENOENT`.
The parent directory option for *mkdir* would not manage the home directory with the appropriate owner, group, and mode, so it's not exactly useful for this either.
The *ceph-common.postinst* uses a more elaborate way to handle this since it allows debian state overrides, this seems unneccessary here since the user is only created exactly once if it does not exist.

Signed-off-by: benaryorg <binary@benary.org>